New organization to Server-Side.
Divided "queryController.ts" in "albumsController.ts", "tracksController.ts" and "artistsController.ts".
Divided "queryRoutes.ts" in "albumController.ts", "trackRoutes.ts" and "artistController.ts".
