+++ server/controllers/publicController.ts
@@ -1,3 +1,4 @@
+// Types
 import { Request, Response } from 'express'
 
 export async function index(_req: Request, res: Response) {
